Add Guid, Date, and DateTime APIs and tests.
